refactor: extract doctor check logic from command
