+++ papis/commands/update.py
@@ -402,7 +402,7 @@ def run(
     [document.pop(k) for k in to_drop]
 
     if auto_doctor:
-        from papis.commands.doctor import fix_errors
+        from papis.doctor import fix_errors
 
         logger.info(
             "Running doctor auto-fixers on document: '%s'.", describe(document),
